Added equals and hashCode methods to PathyObject and subclasses. Action didn't need it adding explicitly.
Finally trying to get junctions to work like more than action-less nodes. The multi-rule internal linking is gonna take some work

Links no longer allow null endpoints.

LinkPairs will be used to perform the multi-rule internal linking for Junctions as they need to equate to true as long as the links in it are the same, regardless of order.