Nodes and Junctions now inherit from PathyPlace. PathyPlace is an abstract subclass of PathyObject, this makes checks involing link endpoints a lot easier. PathyPlace deals with knowing what Links the Node or Junction is associated with, again making a lot of checks a LOT cheaper.
Due to the inheritance change, Link's constructors just became a LOT simpler.

Added a new one param query to the grammar and fixed the definiton of MoveEntity

One param queries have been written. Some are rather meaty, so they are likely going to be factored out into their own methods in the future.

Fixed typo (which could have caused a ClassCastException) in visitSetLink1W()

NOTE: No functionality of Pathy has actually been tested at this point. That happens a bit later, then I go on a bug hunt as there WILL be bugs. If I denied that, I wouldn't be a proper programmer.