Tidying up before the bug bash.
Unused includes are gone.

Path finding will not be implemented at this time as it was not required for the assignmetn grade, and the other project I'm using Pathy on doesn't need the functionality.

Any functionality requiring that functionality will return "ERROR: Not implemented in this version of Pathy"

StackElement was removed as it was never used. It was originally going to do the job that PathyObject ended up doing, as a catch-all type to store in the world Dictionary.