Fixed an issue where a PathyPlace's internal connections HashSet wasn't returning the correct result from contains()
PathyPlace's isConnected() now tests for the connection of a specific Link. hasConnections() now does the job of returning if it has any at all.

Link and Junction Direction now works and is declared with a set of meaningful string values instead of integers.

SetJunctDir is now implemented, as it had somehow been missed.

boolean parameters have been removed from the grammar as they weren't used